+++ src/shared/components/palette/case-flag/components/select-flag-type/select-flag-type.component.ts
@@ -27,7 +27,7 @@ export class SelectFlagTypeComponent implements OnInit {
 
   public get caseFlagWizardStepTitle(): typeof CaseFlagWizardStepTitle {
     return CaseFlagWizardStepTitle
-  };
+  }
 
   public ngOnInit(): void {
     this.flagTypes = this.getFlagTypes();